Separate the old implementation in preparation to the new one
