+++ src/main/java/net/hamnaberg/jsonstat/v1/Category.java
@@ -1,4 +1,4 @@
-package net.hamnaberg.jsonstat;
+package net.hamnaberg.jsonstat.v1;
