+++ src/main/java/net/hamnaberg/jsonstat/v1/table/Renderer.java
@@ -1,4 +1,4 @@
-package net.hamnaberg.jsonstat.table;
+package net.hamnaberg.jsonstat.v1.table;
 
 public interface Renderer<A> {
     A render(Table table);
